switched dashboard and home files, switched /dashboard to /home
