+++ src/components/welcome/LoginForm.js
@@ -107,7 +107,7 @@ const Login = props =>{
         if(loginTeacher){
           console.log(loginTeacher.accessToken)
           setAcsToken(loginTeacher.accessToken)
-          return navigate("/dashboard")
+          return navigate("/home")
         }
       },
       onError(){
